Provide maximum product on shopping cart product update.
In #1163, an issue with updating the cart was identified.

The solution to that issue (#1164) prevented assigning the maximum product
to the cart on update and instead would result in the last acceptable
quantity remaining in the cart instead of going "out-of-bounds".

This corrects the issue(s) discussed in those two threads by minorly
rearranging the actionUpdateProduct code to perform the cart decrease actions
before any cart increase actions and specifically for mixed product as
that is a situation where two or more cart products can/do depend on
each other.

Fixes #1373