Add sanitizer rules for admin/manufacturers.php
The admin/manufacturers.php file includes $_POST data that is similar to generic product; however, there have been no sanitization rules specifically applied which causes all information to be sanitized (potentially overly so) including the url, the name, and the file path related to the image.  This closes that loop at least for the data provided by admin/manufacturers.php.  The rules are applied broadly so that other pages may edit/modify that data if so chosen instead of applying the filtering to just the manufacturers page through the `addComplexSanitization` method.