Cart math before rounding
During certain conditions some of which are described in the ZC forum at:
https://www.zen-cart.com/showthread.php?209786-Shopping-cart-Subtotal-not-equal-Checkout-Subtotal-(rounding-error)-with-GST-VAT
and
https://www.zen-cart.com/showthread.php?221816-Sub-Total-amount-out-by-1-cent

Basically that there are two ways to address the fact that sub-total +
other factors do not total correctly.

In the other solution (to also be posted), the results are provided where
the individual price of the product is sustained regardless of the quantity
of product purchased.  A draw-back to this method is that the prices are
rounded before multiplying and in many cases rounding occurs in multiple
locations while working towards a final solution.  This works best if the
pre-tax price is entered instead of calculated.

In this method, rounding occurs only at the point of display/export/storage
There is also the benefit that currency conversion occurs before any
rounding which maintains all of the values along the way to only round at
the last action/display.  Further, this method doesn't prevent items from
having an individual cost less than the the currency's number of decimals.
In this way entering the after tax price of the product is still respected
and the tax/sub-total of the product is calculated.

There have basically been two conversations about how to "properly" run
these calculations.  Both are being provided for review/determination to
incorporate.  Even if the consideration is to 1) offer a switch to be able
to choose one 2) apply one for product prices that fall into a particular
range, etc...

As a closer, there are some things to clean up; however, the purpose of this
presentation is to discuss the issue and/or determine if there is a "proper"
method of calculation.