Restore currency conversion for 0 decimal length currencies.
A few years ago, PayPal required adjustment to address currencies that do
not have/use a decimal.  At that time, PayPal was updated to first multiply
the store's price by the destination currency and then to cast the value
as an integer (always round down).  This is reasonable (a small amount
is lost by the always down rounding, but it is 1 unit compared to possibly
many, for example currently the rate exchange is about 1 to 116.75 for yen
so 1 yen is less than a 0.01, but when/if the original currency is in the
100s or more, such rounding can be significant over many purchases.

This modification offers an optional parameter to always truncate if nothing
is provided; however, the admin currencies area has been modified to not
perform this truncation when downloading, but instead to happen as has
become the norm whenever the optional parameter is not provided.  This way
existing plugins will continue to operate as they have and if desired the
additional parameter could be added to permit updating the currency to not
have it pre-maturely truncated.