Recognize that site can be all SSL by https: in HTTP_SERVER
Authorize.net AIM functionality is supported only when the site supports SSL communication with the client. The suggestion for setup in the ZC forum for a site that is to be hosted entirely by SSL is to modify the includes/configure.php file such that 1) HTTP_SERVER begins with https: and to change ENABLE_SSL to false. In the current and past versions of ZC to include version 1.5.5 and below, authorize.net AIM is disabled to the customer if ENABLE_SSL was false (not true is the actual test). This commit adds a check that if HTTP_SERVER doesn't begin with https: then if the remaining/existing checks evaluate as true that the payment module would be set to false/be disabled. This continues to support the store front side of offering payment communication securely as offered by current SSL. 

This solution was proposed in the ZC forum: https://www.zen-cart.com/showthread.php?222122-Help!-Authorize-net-(AIM)-has-disappeared-from-checkout!&p=1328484#post1328484