Update product header to include the buy_now_qty
Adding the result of the zen_get_buy_now_qty into the product header removes the processing (and potential modifications necessary to support alternate processes) from the template page.  This allows the template to present the information that is needed/expected without relying on additional processing or dependency on internal functions.

This has also been considered as a way to solve an issue that can occur when using the product option type Grid as offered by Products attributes grid where specifically the use of units or minimum quantities are intended to only impact the individual attribute selections not the overall "product".