Support ISO-8601 style date-of-birth date
Support date of birth collection from "cell phones" as suggested in the forum: https://www.zen-cart.com/showthread.php?222560-New-customer-using-smartphone-Date-of-Birth-info-not-accepted&p=1333443#post1333443

Source of the issue appears to be by the use of `<input type="date" />` in the template file which returns an ISO-8601 styled date.  This date style may not conform with the date format "recommended" to the visitor.  This change does not prevent a guest from entering the data in the ISO-8601 (or expanded as provided by this code) when typing in the number, but it does support the use of "javascript" style date pickers to provide the date in a format understandable for further processing.

An ISO-8601 specifically formatted date would be one that has a dash (-) only between dates and does not include the additional characters supported by the incorporated regex.  To make this regex only recognize ISO-8601 "alternate" formats, change `(|-|\/)` to `(-)`, otherwise the existing regex permits all numbers to be "pressed" together, to have a dash (-) or to have a forward slash (/); however, it is expected that the same symbol/method is used for each separator (or lack of), otherwise the date is considered incorrectly entered.