Substitute function for repetative code
Added the function setCurrentArrayName which is passed by reference the current array name that is to be updated (can be NULL or undefined) and also passed the parameter name that is to be added.

If the current array name does not have any data (empty quotes or not defined), then the returned array name is the parameter name.  If the array name has already been assigned, then the parameter name is appended as an array like name to the existing current array name and the result is returned back to be used/processed.

This was done to simplify expansion on potential future filters, to offer controlled building of the array name in a central location, and to minimize the changes necessary (that could also go wrong) to support future modifications.  The first parameter was set to be received by reference to support the first parameter not being defined and was found to be functional by online PHP testing for PHP 5.2.17 up to PHP 7.1.0.