Rework setCurrentArrayName implementation and use
In discussion with @zcwilt and @drbyte, it became almost painfully obvious that we are all trying to accomplish the same task, have received constructive criticism not only about the direction of this modification but also good programming practices as it relates specifically to ZC as well as general PHP implementation.  As a result of that discussion and yet additional review of the code, identified that 1) there remained areas where setCurrentArrayName should have been used but wasn't (result of hard-coding that arrayName was known to be defined and that a new key was being added), 2) attempting to pass a specific `$currentArrayName` was unnecessary at least in the default filters (custom filters yet to be further investigated), 3) $this->arrayName could always be set as necessary around the operation to be performed to support further building. 4) There was a bit of redundant code in the more complex functions that could be removed/relocated.

Part of this rewrite did identify that the new function `setCurrentArrayName` does not accept, deal with nor handle the situation of if the $parameterName is an array or if $this->arrayName is or becomes an array.  Notes were left to offer others the opportunity to answer the questions that prevented taking a direction with either of these conditions.  Ie. should all elements be handled equivalently, would the size of the arrays need to be equivalent (take five elements and apply each to a list of five elements matching the key or position within each array or if one is an array the other not, etc...), or applied to the last element, etc... I have not yet conceived how this would occur or what it would address.

As part of the discussion did elude to more object oriented programming, it does appear that there are some helper functions that have not been used or perhaps have been avoided that could be incorporated to improve on the OOP aspect. This has not been adjusted in this commit to try to keep the change as pure as necessary to support the addition of the simple array and changes needed to support the associated processing modifications.