Replace (isset && > 0) with !empty
In developer/owner review, comment was made that the combination function of: `isset($var) && $var > 0` could be shortened to `!empty($var)`

Note that other design considerations are that `$_GET[$var]` variable is not to support arrays and would be expected to be sanitized as such prior to code usage, therefore, of the `empty()` considerations, a value of 0, "0", 0.0, '', FALSE, or NULL would be the only ones that would result in a true condition for `empty()` which would be a false condition for `!empty()`. Conversely anything other than those values would result in executing the code that follows and that the variable `$_GET[$var]` would *not* be an array as provided on the URI of any size. 

This commit performs that revision to support strict code operation.

Also included is a removal of single quotes around integer values (reduces database operations of converting the then string integer back to an integer) and using all capital letters for mySql "commands".