Clean up upcoming_products.php module
Replace (!isset && >0) with !empty,
Replace `!isset || == '0'` with empty,
Capitalize mySql query commands,
Use ZC 1.5.5 style of database iterator (replace while (!$var->EOF) with foreach)
In developer/owner review, comment was made that the combination function of: `isset($var) && $var > 0` could be shortened to `!empty($var)`

Note that other design considerations are that `$_GET[$var]` variable is not to support arrays and would be expected to be sanitized as such prior to code usage, therefore, of the `empty()` considerations, a value of 0, "0", 0.0, '', FALSE, or NULL would be the only ones that would result in a true condition for `empty()` which would be a false condition for `!empty()`. Conversely anything other than those values would result in executing the code that follows and that the variable `$_GET[$var]` would *not* be an array as provided on the URI of any size. 

This commit performs that revision to support strict code operation.