Again rework array data
Incorporated possibility that $this->fields may be set in queryFactoryResult if it has been edited or overridden so that can either use the memory/process cheap or a little more effort to loop through Execution of a query, also, for Move method, incorporated a "use the same method as before" process unless that is overridden as well when Move is explicitly called from outside of the queryFactoryResult class.  This way the Move method when called on a particular result set will always update the same set of data that it had previously updated and not cause a mixture of data being updated (again unless overridden/edited between calls to the Move method).