Modify onClick for tr to applicable td
It has been seen on some apple/macintosh devices that combining
onclick javascript actions in a parent "group" such as the tr of a table
with clickable links within a child (td) causes an override of the onclick
javascript action to take place. Complete removal of the onclick event
resolves the one issue, but that removes the "easy button" that was
created.

This commit pulls the javascript onclick event into the necessary/associated
area that does not otherwise have a direct link.  Meaning the "easy button"
should remain available if all edits have been correct.