Correct controlling logic.
Previous operation had returned from the function when a method
did not provide any quotes.  That was modified to remain in
the function and allow other function steps to occur; however,
the issue to be addressed was to stop processing the one method
if that method had no quotes and continue processing the next
method.

Based on the conversation at the original pull #2078, the
OP has agreed that it should be `continue` or something that
provides that operation.  Now that I've actually looked at what
is going on and the potential resulting problem if something
were not done, I can see where departing the function overall
at that point certainly was an issue.