Display shopping_cart messages one time.
Identified that when a product has attributes and is mixed that if one of the criteria is not met (minimum, maximum, units) that the warning message is produced for each instance rather than one time per main product.  If the variable `$this->flag_duplicate_quantity_msgs_set['keep']` had been set somewhere in this process then the display would be performed one time.  Moved the definition of the array `$this->flag_duplicate_quantity_msgs_set` to outside of the loop that supports redisplaying the message.

This appears to be a change in operation.  I believe originally it was added in so that one could choose to allow the repetitive display or disable the repetition and show only once when using other pre-checks in the extra_cart_actions folder, but still seems like it may be of value to show only once in standard operation.