Align data type with other existing data.
Okay, not a great description, but here's what is/was going on.
Previously `STORE_OWNER_EMAIL_ADDRESS` was being used as the field
to represent the name of the `to_name`; however, if that did not
include both parts of a name and an email address then the send
would fail.  So to correct the issue, the use of `STORE_OWNER_EMAIL_ADDRESS`
was replaced in the `to_name` area to use `STORE_NAME`. While that
generalizes the `to_name` field, there was one case of zen_mail
where the value was $oldData['admin_email'] that was changed
to `STORE_NAME`. The line appears to use `$oldData` for
referential information, so it only seems appropriate that if
it exists, that the same `$oldData` admin_name be used.

This converts the use of `STORE_NAME` to `$oldData['admin_name']`
in one case where other `$oldData` is used.