Adjust added product to max of max and units
When adding a product to the cart, where the addition of the item results in more product in the cart than allowed, the resulting quantity was set to the maximum quantity in stock followed by the maximum quantity allowed for the product and/or mix of attributes. The originally calculated maximum may not align with units where this commit corrects the maximum to account for the product's units.  This takes out one factor of a customers concern about what it takes to not exceed the maximum and still meet the units.  It will though likely lead to confusion about why they can not add more especially if there is some sort of indication of what the maximum quantity of product is that is in stock when units are a factor.  

This modification applies only to adding product to the cart, a separate commit is required to address other aspects such as updating the cart or other such `action` sections. Ultimately this prevents adding a quantity that violates the units when attempting to add the maximum allowed/available quantity.