v158 Permit email to be sent to follow-on recipient(s)
This allows continuing the loop of possibly sending an email to the identified recipient(s) even if the most "recent" recipient is identified as not to receive the email. For those intending to prevent the subject email from being sent, they would need to retain that ban until at least leaving the `foreach` loop.