Update upgrade path for configuration_description
While searching settings and reading the configuration description presented in the admin, identified that the expressed default did not match the current setting (default) and did not match any of the available options for that configuration.  This updates the upgrade path to include the current default value for this `configuration_description`.