Update init_sefu.php
Provide early return.
Properly implement for loop incrementing by value of 2.
Replace array sizeof with count.
Apply spacing in variable assignment. 
Replace comment guidance for more information from the old wiki link to the current docs location of the same information. 
Result of the above is to shift all code to the left and remove nested if and other code.