Correct isXxxDir if just current directory path (__DIR__) given
Expectation of each DIR_FS_ADMIN and DIR_FS_CATALOG is that they
end with a DIRECTORY_SEPARATOR (`/`). If the complete path/directory
excluding the filename where the path is at the root, then this change
ensures to use the "shortest" path representative of the associated folder
to answer the boolean question.